requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
requests==2.31.0
urllib3==1.26.15
aiohttp==3.9.1